requirements.txt
nose==1.3.1
paramiko==1.10.1
pycrypto==2.6.1
#python-apt===0.9.3.5ubuntu1
#python-debian===0.1.21-nmu2ubuntu2
Twisted-Core==13.2.0
